+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Correctly send and display strikethrough text via the del-tag
 - Fix room settings crashing on opening when using Android
 - Whitespace is stripped when searching for mxids
+- Fix focus handling in chat/group creation user list
 
 ### Security